requirements.txt
torch>=2.1.0
transformers>=4.46.0
peft>=0.13.0
bitsandbytes>=0.44.0
accelerate>=1.1.0
sqlparse>=0.5.0
func-timeout>=4.3.0
pyyaml>=6.0
tqdm>=4.65.0
datasets>=2.14.0
matplotlib>=3.7.0
